map service for converting to DTO object & add save,findbyid methods

+++ src/main/java/ru/itmo/aquacourseproject/DTOmodel/LocationEntityDto.java
@@ -8,7 +8,7 @@ import java.math.BigInteger;
 @Data
 public class LocationEntityDto implements Serializable {
     private final int id;
-    private final BigInteger x;
-    private final BigInteger y;
-    private final BigInteger z;
+    private final Double x;
+    private final Double y;
+    private final Double z;
 }
